MirroringConfigurationHelper
